+++ src/api_client.rs
@@ -1,15 +1,16 @@
+use std::sync::Arc;
+
 use super::activities_api::ActivitiesApi;
 use super::configuration::Configuration;
-use std::rc::Rc;
 
 pub struct ApiClient {
-    pub configuration: Rc<Configuration>,
+    pub configuration: Arc<Configuration>,
     pub activities_api: Box<ActivitiesApi>,
 }
 
 impl ApiClient {
     pub fn new(configuration: Configuration) -> ApiClient {
-        let rc = Rc::new(configuration);
+        let rc = Arc::new(configuration);
 
         ApiClient {
             configuration: rc.clone(),